+++ CHANGELOG.md
@@ -8,9 +8,16 @@ The date format used is DD/MM/YYYY.
 
 ## [[1.2.0]](https://github.com/willymateo/titi-backend-sql/pull/7) - 04/01/2022
 
+### Added
+
+- Update adventure
+- Delete adventure
+
 ### Changed
 
 - in the GET /adventures only send adventures with publisher different that in the token
+- Change PUT to PATCH method of /me/account route.
+- Update dependencies
 
 ### Removed